[Buffer] Added Makefile
Added Makefile that, through the target 'all', calls iverilog with the circular_buffer.v file and its testbench file; then it runs the created files (a.out) to execute the simulation and finally calls gtkwave with the vcd file.
The Makefile has also another target 'clean' that simply remove the a.out and out.vcd from the current folder.