[UTILS] Created a package for utility functions
The Utils package contains utility functions, such as the one that calculates the ceiling of the base 2 logarithm, which was already used in more than one module, and will supposedly be used in many others in the future.
The function was extracted from the Circular Buffer module and put inside the Utils package.