[BUFFER] Added testbench
Created a basic testbench module that instantiates as a DUT the circular buffer and tests it with some combinations of read/write (and random data) input signals. It covers also some border situations such as write when full and read when empty. Finally it executes random read/write operations, that sometimes can be simultaneous.