[RCUNIT] Substituted integer ports and nets with logic arrays
In the Route Computation Unit module and the related testbench, the integer values representing the destination coordinates (and the offset values with reference to the current position, consequently) have been substituted by using a properly-sized logic array (i.e., of the same size as the destination coordinates encoded in the head flit structure).
Also, better usage of the common parameters has been implemented.