[INTERFACE] First version interface between Input Port and VC Allocator
Added an interface connecting an Input Port to the Virtual Channel Allocator, so that the VC Allocator can send to the Input Port the Virtual Channels to assign at the downstream router to the currently stored flits.
For now, only a connection directed from the VC Allocator to the Input Port is provided; in the future, further ports may be added to implement other functionalities.