[INPUTPORT] Implemented the Input Port module
The Input Port module is composed of one Input Buffer for each Virtual Channel and one Route Computation Unit.
The module is made of combinational logic only (other than the contained Input Buffers, which are finite state machines), and is interfaced with the Crossbar, Virtual Channel Allocator and Switch Allocator modules.
The module has yet to be properly tested on a testbench.