[INBUTBUFFER] First, basic, version of the testbench
A first basic version of the tb for the IB has been developed. The tb simply feed the buffer with subsequent head, body, tail flit. 
The next step of enhancement should include a  flit queue, to automatize controls, and some corner cases check to verify the IB works as expected.