[INPUTBUFFER] Fixed testbench
The tb of the IB now properly inserts and removes flits from the queue. While reading them from both queue and buffer, it also performs an equivalence check (and if it fails, the simulation stops). For this reason, the structure of the read and write operations has been modified with respect to the previous version. 
Finally, the create_flit task now  correctly creates the different types of flit according to its label, differentiating between the ones that are HEAD and the ones that are not (BODY and TAIL).