[INPUTBUFFER] Upgraded testbench
The tb of the IB now performs reads in between the insertion of the various flits of each packet. Also the function that checks the correctness of the flit has been upgraded and features the change of the vc_id of each flit. The vales of the flits are also now influenced by a number that only increases during the program execution. This factor guarantees the uniqueness of each flit, that was not guaranteed in the previous version of the testbench.