[INPUTBUFFER] Added VC change check to testbench
The testbench now checks the VC change. The VCs are inserted manually into the testbench in order to test that the input buffer works correctly. The changes are managed by the queue that changes the vc_id with the expected value and then check at the read stage that the change has been done correctly.