[INPUTPORT] Added integer for operation identifier
In the tb of the inputport an integer number identifier has been added in order to keep track of the unique number of the operations which are performed (i.e. read or write of a single flit) with the purpose to uniquely identify each flit inserted into the input port and read out of the crossbar.