[INPUTBUFFER] Restricted Write operations based on current state
Restricted the Write operations in the Input Buffer module, depending on the current state of the FSM; when in Idle state only one Head flit can be written in the buffer, while in VA and SA only Body flits and one Tail flit can be written.
Proper checks also guarantee that, after writing in the Buffer the Tail flit of the current packet, no additional flit is accepted until the FSM has returned to the Idle state.
Also, a minor fix was applied to the read_cmd logic, without changing its behavior at all.