[ALLOCATOR] Implemented Separable Input-First Allocator
A module for the allocation of multiple shared resources between multiple agents has been implemented by using the Separable Input-First Allocation structure.
This module is supposed to be later used as the foundation of both Virtual Channel and Switch Allocators, and has to be properly tested by means of a testbench.