[VCALLOCATOR] Added first version of Virtual Channel Allocator
A first implementation of the Virtual Channel Allocator module has been added, by wrapping a Separable Input-First Allocator module with proper combinational and sequential logic.
The module is synthesizable on Vivado without any warning, but has yet to be tested.
Also, inputs and outputs of the module will later be adjusted to better integrate through interfaces with Input Port modules.