[INPUTBUFFER] Adapted the module for VC Allocation
Outputs were added to the module to enable communicating to the Upstream VC Allocator when the Input Buffer can be allocated to a new packet and requesting to the VC Allocator of the local router for the allocation of a Downstream Virtual Channel.
The new functionalities will later be properly tested in the corresponding testbench.