[INPUTPORT] Adapted the module after VA changes
A signal has been added to communicate to the upstream router the availability of VCs inside the Input Port to be allocated.
Furthermore, the next hops computed by the Route Computation Unit have been tweaked to be propagated to both the Switch Allocator and Virtual Channel Allocator interfaces.
These latest changes will later be properly tested.