[SWITCHALLOCATOR] Added first version of the Switch Allocator module
A first version of the Switch Allocator module has been implemented, by making use of the previously developed Round-Robin Arbiter and Separable Input-First Allocator modules.
In the upcoming versions, the module is expected to make use of interfaces towards the interacting modules (Input Block and Crossbar), and the correct behavior of the implementation will have to be proved by means of a testbench.