[INPUTPORT] Fixed Testbench
The main task of the tb, insert_packet, has been modified in order to support new parameters; this allows specific settings of the tests to execute. For this reason, the tasks for reading and writing flits have been modified too.
Moreover, the need of the extra-cycle in advance before the packet insertion has been removed. 
Finally the task for the read of the entire packet has been deleted.
Some corner-case tests will be provided later.