[VCALLOCATOR] Added controls to the testbench
The testbench now includes an internal implementation of a separable input first allocator and simulates the behaviour of the module. After the output from the simulation is computed it is then compared with the output of the real module. If there's an error a message will be displayed and the simulation is interrupted.