[INPUTBUFFER] Bugfix: checks for single flit packet
In the FSM of the IB, the final transition from SA to IDLE did not consider the case of a single flit packet and in this case the FSM didn't correctly return into IDLE state.
Moreover, in VA state it was previuosly possible to receive more than one HEADTAIL flit without properly signal the wrong insertion (i.e. raising error flag)