[CROSSBAR] Added interfaces support in testbench
The testbench of the crossbar now properly supports the interfaces introduced between the crossbar and both the input block and the switch allocator using a mock module.
Moreover, two warnings related to a return instruction (now substituted with a break one) and the definition of the input parameter of the function 'corresp_output' have been resolved.