[BUFFER] Removed Pipeline Depth parameter, introduced On/Off Latency one
The On/Off flow control was configured with a wrong parameter, i.e., the depth of the pipeline, when it is instead more appropriate to set as its propagation latency a value of 2; instead of a parameter, this was added to the Circular Buffer module as a localparam.
The parameter has also been removed from all the modules enclosing a Circular Buffer (up to the Router), and their corresponding testbenches.