[ALLOCATORS+ROUTER] Discarded modules: performance and area concerns
The modules for the Virtual Channels and Switch allocation have been deemed as insufficiently performing from both the area and critical path length points of view.
Also the speculative implementation will later be reworked, to make it simpler and more efficient than the previous one.
Therefore, we are discarding the Separable Input-First Allocator, the VC Allocators, the Switch Allocators and the Router modules.