[SWITCHALLOCATOR] Added new Switch Allocator module
The Switch Allocator module contains a Separable Input-First Allocator; it translates the requests for the Crossbar traversal from the Input Buffers into the requests to be processed from the SIF Allocator, and from its grants drives the Buffers read and the Crossbar switch.
The new SIF Allocator implementation has definitely simplified the Switch Allocator logic, making it a smaller and faster module.