[IB, XBAR, VA, SA] Created basic testbench to check interconnections
A basic testbench containing all the basic blocks of the router (i.e. input block, crossbar, virtual channel allocator and switch allocator) has been created with the purpose of simply checking that all the interconnections between the different modules work properly. The testbench executes just the insertion of a single packet composed of 4 flits; the outcome of this test should be checked in a graphical way, since any checking tasks is provided.
A more complete and structured testbench, that will be provided later, will cover the entire router module using its external interfaces.