[INPUTBUFFER] Bugfix: switch request signal
The signal for the switch request, namely switch_request_o,  was previously asserted every time the buffer was in SA state; this was wrong because it did not consider that the internal buffer could possibly be empty.  Now the signal is raised only when the internal buffer is not empty.