[ROUTER] Added test case for long packet
The tb now exectues also tests with packets whose length exceed the buffer size. 
In order to properly test this case, in the flit insertion task has been introduced another boolean condition that checks the value of the on/off flag, of the specific port and vc identifier, before effectively proceed with the insertion operation.